+++ tests/fixtures.c
@@ -181,24 +181,28 @@ FIXTURE_FN(471141295)
     FIXTURE_STRING("dude");
 }
 
-static int value_for_510017722 = 1;
-FIXTURE_FN(510017722)
+static int value_for_4284892664 = 1;
+FIXTURE_FN(4284892664)
 {
-    // "function (options) {\n          if( typeof value === 'undefined' ) { value = 1; } return options.fn({value: 'bar'}, {blockParams: options.fn.blockParams === 1 ? [value++, value++] : undefined});\n        }"
+    // "function (options) {\n          if( typeof value === 'undefined' ) { value = 1; } return options.fn({value: 'bar'}, {blockParams: options.fn.blockParams === 1 ? [global.value++, global.value++] : undefined});\n        }"
     HANDLEBARS_VALUE_DECL(context);
     HANDLEBARS_VALUE_DECL(bp1);
     HANDLEBARS_VALUE_DECL(bp2);
     HANDLEBARS_VALUE_DECL(block_params);
+    struct handlebars_value * effective_block_params = NULL;
 
     handlebars_value_init_json_string(CONTEXT, context, "{\"value\": \"bar\"}");
-    handlebars_value_integer(bp1, value_for_510017722++);
-    handlebars_value_integer(bp2, value_for_510017722++);
+    if( options->program_block_params == 1 ) {
+        handlebars_value_integer(bp1, value_for_4284892664++);
+        handlebars_value_integer(bp2, value_for_4284892664++);
 
-    handlebars_value_array(block_params, handlebars_stack_ctor(CONTEXT, 2));
-    handlebars_value_array_push(block_params, bp1); // @TODO ignoring return value - should probably make a handlebars_value_push()
-    handlebars_value_array_push(block_params, bp2); // @TODO ignoring return value - should probably make a handlebars_value_push()
+        handlebars_value_array(block_params, handlebars_stack_ctor(CONTEXT, 2));
+        handlebars_value_array_push(block_params, bp1); // @TODO ignoring return value - should probably make a handlebars_value_push()
+        handlebars_value_array_push(block_params, bp2); // @TODO ignoring return value - should probably make a handlebars_value_push()
+        effective_block_params = block_params;
+    }
 
-    struct handlebars_string * tmp = handlebars_vm_execute_program_ex(vm, options->program, context, NULL, block_params);
+    struct handlebars_string * tmp = handlebars_vm_execute_program_ex(vm, options->program, context, NULL, effective_block_params);
     handlebars_value_str(rv, tmp);
 
     HANDLEBARS_VALUE_UNDECL(block_params);
@@ -1695,7 +1699,8 @@ static void convert_value_to_fixture(struct handlebars_value * value)
         FIXTURE_CASE(461441956);
         FIXTURE_CASE(464915369);
         FIXTURE_CASE(471141295);
-        FIXTURE_CASE(510017722);
+        FIXTURE_CASE_ALIAS(510017722, 4284892664);
+        FIXTURE_CASE(4284892664);
         FIXTURE_CASE(585442881);
         FIXTURE_CASE(620640779);
         FIXTURE_CASE(620828131);